requirements.txt
octokit==2.0.6
pandas==1.5.3
matplotlib==3.8.2
pytz==2023.3
python-dotenv==1.0.0